industrial safety scanner/requirements.txt
wincertstore==0.2
boto3==1.16.51
botocore==1.19.51
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
DateTime==4.3
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
line-bot-sdk==1.18.0
MarkupSafe==1.1.1
numpy==1.19.5
opencv-python-headless==4.2.0.32
pyDes==2.0.1
python-dateutil==2.8.1
pytz==2020.5
qrcode==6.1
requests==2.25.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
zope.interface==5.2.0